+++ R/osm_query_changesets.R
@@ -2,7 +2,10 @@
 #'
 #' This is an API method for querying changesets. It supports querying by different criteria.
 #'
-#' @param bbox Find changesets within the given bounding box coordinates (`left,bottom,right,top`).
+#' @param bbox Find changesets within the given bounding box coordinates (`left,bottom,right,top`). It can be specified
+#'   by a character, matrix, vector, `bbox` object from \pkg{sf}, a `SpatExtent` from \pkg{terra}. Unnamed vectors and
+#'   matrices will be sorted appropriately and must merely be in the order (`x`, `y`, `x`, `y`) or `x` in the first
+#'   column and `y` in the second column.
 #' @param user Find changesets by the user with the given user id (numeric) or display name (character).
 #' @param time Find changesets **closed** after this date and time. See details for the valid formats.
 #' @param time_2 find changesets that were **closed** after `time` and **created** before `time_2`. In other words, any
@@ -132,7 +135,7 @@ osm_query_changesets <- function(bbox, user, time, time_2, from, to, open, close
   if (missing(bbox)) {
     bbox <- NULL
   } else {
-    bbox <- paste(bbox, collapse = ",")
+    bbox <- bbox_to_string(bbox)
   }
 
   if (missing(user)) {