+++ NEWS.md
@@ -3,6 +3,8 @@
 * Update documentation and code for server-side changes documented in OSMWikiVersion
   [2878437 -> 2940426](https://wiki.openstreetmap.org/w/index.php?title=API_v0.6&diff=2940426&oldid=2878437) (#72).
   * Add new function `osm_search_comment_changeset_discussion()`.
+* `bbox` parameters now accepts more formats, including character, matrix, vector, `bbox` object from `sf` package, or a
+  `SpatExtent` from `terra` package (#73).
 
 # osmapiR 0.2.4